Removed cluttered docs and added quickstart
